Show unsigned request status code
Rather than show a blank image, which has a default broken look (that's
sort of the point of the example, but may cause confusion). More
practically, the output of the bad `Image` display does not contain an
alt text attribute, causing an accessibility error. Programatically
adding some default alt text when rendering will actually show the alt
text value, since the image isn't displayed. This is doubly confusing.
In order to pass accessibility tests and show clearly that the request
failed, I switched to showing the status code of an explicit request to
that URL instead of rendering a "broken" image.