Update link to Example NAIP notebook
Update the NAIP notebook link to the html catalog notebook tab. The
previous relative path resolves correctly in the Hub but not when
rendered in the docs site (a 404). The absolute URL to the same (rendered)
content will work for both, and leaves the "Launch" button to get it to
open the actual notebook in the Hub.