Restructure reading-stac quickstart
This rewrites the reading-stac quickstart. It emphasizes using the
search API and defers details on browsing / walking collections until
later.

It also demonstrates using STAC metadata as data (geojson features),
via geopandas.

Finally, it fixes the issue reported in
https://github.com/microsoft/PlanetaryComputer/discussions/53 by
avoiding the use of the "thumbnail" asset on Landsat, which aren't
present after 2021.